streamlit/requirements.txt
streamlit
kafka-python
pandas
minio
python-dotenv